# cmd/browser-agent/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Kaboom MCP Server",
    "description": "Browser observability for AI agents. Streams console logs, network errors, and user actions from Chrome to Claude Code, Cursor, and Copilot via MCP protocol.",
    "version": "0.7.12",
    "license": {
      "name": "MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:{port}",
      "description": "Local Kaboom server",
      "variables": {
        "port": {
          "default": "7890",
          "description": "Server port (configurable via --port flag)"
        }
      }
    }
  ],
...
  "paths": {
...
    "/websocket-status": {
      "get": {
        "tags": [
          "Data Read"
        ],
        "summary": "Get WebSocket connection status",
        "description": "Returns active and recently-closed WebSocket connections with message rates and sampling info. Used by MCP clients via observe(what: 'websocket_status') to understand which WebSocket connections are active on the page, their throughput, and whether sampling is being applied to high-volume connections.",
        "operationId": "getWebSocketStatus",
        "security": [
          {
            "extensionClient": []
          }
        ],
        "responses": {
          "200": {
            "description": "Active and closed WebSocket connections with throughput stats",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WebSocketStatusResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "extensionClient": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Kaboom-Client",
        "description": "Extension client identifier. Must match pattern 'kaboom-extension-*'. Required for extension-only endpoints."
      }
    },
    "schemas": {
...
      "WebSocketStatusResponse": {
        "type": "object",
        "description": "Overview of active and recently-closed WebSocket connections. Includes message rates and sampling info for each connection, used to identify high-throughput connections and sampling behavior.",
        "properties": {
          "connections": {
            "type": "array",
            "description": "Currently active WebSocket connections",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string",
                  "description": "Connection ID"
                },
                "url": {
                  "type": "string",
                  "description": "WebSocket URL"
                },
                "state": {
                  "type": "string",
                  "description": "Connection state (e.g. 'open')"
                },
                "opened_at": {
                  "type": "string",
                  "format": "date-time",
                  "description": "When the connection was opened"
                },
                "duration": {
                  "type": "string",
                  "description": "Human-readable connection duration"
                },
                "message_rate": {
                  "type": "object",
                  "description": "Messages per second statistics"
                },
                "sampling": {
                  "type": "object",
                  "description": "Sampling configuration if applied to this connection"
                }
              }
            }
          },
          "closed": {
            "type": "array",
            "description": "Recently closed WebSocket connections (kept for short-term history)",
            "items": {
              "type": "object"
            }
          }
        }
      },
...
    }
  }
}